UDKPB/requirements.txt
django
torch
transformers
Pillow
opencv-python
numpy
ultralytics
mysqlclient
